Move create_database impl out of lib.rs
